Renamed `PathContainer` to `SegmentingContainer`
And added a detailed description with example usage.
